chieu25/6

+++ database/migrations/2020_05_14_000602_create_bills_table.php
@@ -21,7 +21,8 @@ class CreateBillsTable extends Migration
             $table->integer('price_regulation_id'); 
             $table->integer('payment_month');
             $table->integer('payment_year');
-            $table->decimal('money_to_pay',15, 2); 
+            $table->integer('money_to_pay');
+            //$table->decimal('money_to_pay',15, 2); 
             $table->integer('usage_level_max')->nullable(); // mức sử dụng cao nhất đối với điện và nước
             $table->boolean('paid')->default(0);
